+++ src/WWS/Order/Domain/OrderItems/OrderItem.cs
@@ -8,7 +8,7 @@ namespace Domain.OrderItems
 {
     public class OrderItem
     {
-        public OrderItem(long id, long orderId, long itemId, string name, float price, int count)
+        public OrderItem(long id, long orderId, long itemId, string name, double price, int count)
         {
             Id = id;
             OrderId = orderId;
@@ -22,7 +22,7 @@ namespace Domain.OrderItems
         public long OrderId { get; set; }
         public long ItemId { get; set; }
         public string Name { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         public int Count { get; set; }
     }
 }
